+++ migrations/20260219165446_add_global_colum_incidents.ts
@@ -6,7 +6,7 @@ export async function up(knex: Knex): Promise<void> {
       table.string("is_global", 15).notNullable().defaultTo("YES");
     });
   }
-  if (!(await knex.schema.hasColumn("maintenances", "is_global"))) {
+  if ((await knex.schema.hasTable("maintenances")) && !(await knex.schema.hasColumn("maintenances", "is_global"))) {
     await knex.schema.table("maintenances", (table) => {
       table.string("is_global", 15).notNullable().defaultTo("YES");
     });
@@ -14,10 +14,14 @@ export async function up(knex: Knex): Promise<void> {
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.schema.table("incidents", (table) => {
-    table.dropColumn("is_global");
-  });
-  await knex.schema.table("maintenances", (table) => {
-    table.dropColumn("is_global");
-  });
+  if (await knex.schema.hasColumn("incidents", "is_global")) {
+    await knex.schema.table("incidents", (table) => {
+      table.dropColumn("is_global");
+    });
+  }
+  if ((await knex.schema.hasTable("maintenances")) && (await knex.schema.hasColumn("maintenances", "is_global"))) {
+    await knex.schema.table("maintenances", (table) => {
+      table.dropColumn("is_global");
+    });
+  }
 }
